moves a star, path smoothing, and queue to separate classes
